- Extract MailMessage class - Move getting "unknown.vm" to catch
